requirements.txt
flask
python-telegram-bot
